refact: TripInfo를 props로 전달하도록 수정
